+++ test/routes/index.test.js
@@ -75,10 +75,6 @@ describe('getRouteHandlers', () => {
     stopImportJob: sinon.stub(),
   };
 
-  const mockAssistantController = {
-    processImportAssistant: sinon.stub(),
-  };
-
   const mockApiKeyController = {
     createApiKey: sinon.stub(),
     deleteApiKey: sinon.stub(),
@@ -119,7 +115,6 @@ describe('getRouteHandlers', () => {
       mockTrigger,
       mockFulfillmentController,
       mockImportController,
-      mockAssistantController,
       mockApiKeyController,
       mockSitesAuditsToggleController,
       mockOpportunitiesController,
@@ -145,7 +140,6 @@ describe('getRouteHandlers', () => {
       'POST /tools/api-keys',
       'GET /tools/api-keys',
       'POST /tools/import/jobs',
-      'POST /tools/import/assistant/prompt',
     );
 
     expect(staticRoutes['GET /configurations']).to.equal(mockConfigurationController.getAll);